exercise 4.20, 4.21 and 4.22 done

+++ part4/backend/controllers/blogs.js
@@ -3,14 +3,6 @@ const Blog = require('../models/blog')
 const User = require('../models/user')
 const jwt = require('jsonwebtoken')
 
-const getTokenFrom = request => {
-    const authorization = request.get('authorization')
-    if (authorization && authorization.startsWith('Bearer ')) {
-        return authorization.replace('Bearer ', '')
-    }
-    return null
-}
-
 blogsRouter.get('/', async (request, response) => {
     const blogs = await Blog.find({}).populate('user', { username: 1, name: 1 })
     response.json(blogs)
@@ -28,11 +20,7 @@ blogsRouter.get('/:id', async (request, response) => {
 blogsRouter.post('/', async (request, response) => {
     const body = request.body
 
-    const decodedToken = jwt.verify(getTokenFrom(request), process.env.SECRET)
-    if (!decodedToken) {
-        return response.status(401).json({ error: 'token invalid' })
-    }
-    const user = await User.findById(decodedToken.id)
+    const user = request.user
 
     if (!body.title || !body.url) {
         return response.status(400).end()
@@ -55,8 +43,18 @@ blogsRouter.post('/', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
-    await Blog.findByIdAndDelete(request.params.id)
-    return response.status(204).end()
+    if (!await Blog.findById(request.params.id)) {
+        return response.status(401).json({ error: 'blogID invalid' })
+    }
+    const user = request.user
+    const blog = await Blog.findById(request.params.id)
+
+    if (blog.user.toString() === user.id) {
+        await Blog.findByIdAndDelete(request.params.id)
+        return response.status(204).end()
+    } else {
+        return response.status(401).json({ error: 'invalid user' })
+    }
 })
 
 blogsRouter.put('/:id', async (request, response) => {
